+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-2.20 (unreleased)
+2.20 (2023-09-04)
 -----------------
 
 - Fixed conversion status update that was not working when using async conversion